refactor: 지역 목록 페이지 리팩토링
