+++ src/app/(local)/local/[region]/loading.tsx
@@ -1,0 +1,7 @@
+import LoadingState from '@/components/LocalList/LoadingState';
+
+function Loading() {
+  return <LoadingState isPending={true} error={null} />;
+}
+
+export default Loading;
