+++ src/components/LocalList/ContentTypeFilter.tsx
@@ -12,10 +12,9 @@ const contentTypes = [
   { id: '39', name: '맛집' },
 ];
 
-const ContentTypeFilter: React.FC<ContentTypeFilterProps> = ({
-  selectedContentType,
-  onContentTypeChange,
-}) => {
+function ContentTypeFilter(props: ContentTypeFilterProps): React.ReactElement {
+  const { selectedContentType, onContentTypeChange } = props;
+
   return (
     <div className="font-pretendard font-semibold text-sm border-solid flex justify-center overflow-x-auto whitespace-nowrap mb-4 px-4">
       {contentTypes.map((type) => (
@@ -33,6 +32,6 @@ const ContentTypeFilter: React.FC<ContentTypeFilterProps> = ({
       ))}
     </div>
   );
-};
+}
 
 export default ContentTypeFilter;
