+++ src/constants/regionImages.ts
@@ -1,0 +1,19 @@
+export const regionImages: { [key: string]: string } = {
+  seoul: '/Seoul.jpg',
+  busan: '/Busan.jpg',
+  incheon: '/Incheon.jpg',
+  daejeon: '/Daejeon.jpg',
+  daegu: '/Daegu.jpg',
+  gwangju: '/Gwangju.jpg',
+  sejong: '/Sejong.jpg',
+  ulsan: '/Ulsan.jpg',
+  jejudo: '/Jeju.jpg',
+  gyeonggido: '/Gyeonggi.jpg',
+  gangwondo: '/Gangwon.jpg',
+  chungcheongbukdo: '/Chungbuk.jpg',
+  chungcheongnamdo: '/Chungnam.jpg',
+  gyeongsangbukdo: '/Gyeongbuk.jpg',
+  gyeongsangnamdo: '/Gyeongnam.jpg',
+  jeollabukdo: '/Jeonbuk.jpg',
+  jeollanamdo: '/Jeonnam.jpg',
+};
